refactor: move duplicated logic into common file
